refactor(android): move IsAndroidTVDevice into the platform unit (2a)
Move the leanback/TV-detection JNI (with its JniCheck helper) verbatim
from SOURCES/TOUCH_INPUT.CPP into LIB386/SYSTEM/ANDROID.CPP. TOUCH_INPUT
now includes <SYSTEM/ANDROID.H> for its call site and drops its <jni.h>
and <SDL3/SDL_system.h> includes. Behavior unchanged — the function
already returned 0 off-Android.
