+++ LIB386/SYSTEM/ANDROID.CPP
@@ -18,8 +18,74 @@
 
 S32 IsAndroidTVDevice(void) {
 #ifdef __ANDROID__
-    /* Real implementation moves here in Phase 2a. */
-    return 0;
+    struct JniCheck {
+        static S32 Pending(JNIEnv *e) {
+            if (e->ExceptionCheck()) {
+                LogPrintf("TouchOverlay: JNI exception in TV detection, "
+                          "overlay disabled\n");
+                e->ExceptionClear();
+                return 1;
+            }
+            return 0;
+        }
+    };
+    JNIEnv *env = (JNIEnv *)SDL_GetAndroidJNIEnv();
+    if (env == NULL)
+        return 0;
+
+    jobject activity = (jobject)SDL_GetAndroidActivity();
+    if (activity == NULL)
+        return 0;
+
+    jclass activityClass = env->GetObjectClass(activity);
+    if (JniCheck::Pending(env) || activityClass == NULL)
+        return 0;
+
+    jmethodID getPM = env->GetMethodID(activityClass, "getPackageManager",
+                                       "()Landroid/content/pm/PackageManager;");
+    if (JniCheck::Pending(env) || getPM == NULL) {
+        env->DeleteLocalRef(activityClass);
+        return 0;
+    }
+
+    jobject pm = env->CallObjectMethod(activity, getPM);
+    if (JniCheck::Pending(env) || pm == NULL) {
+        env->DeleteLocalRef(activityClass);
+        return 0;
+    }
+
+    jclass pmClass = env->GetObjectClass(pm);
+    if (JniCheck::Pending(env) || pmClass == NULL) {
+        env->DeleteLocalRef(pm);
+        env->DeleteLocalRef(activityClass);
+        return 0;
+    }
+
+    jmethodID hasFeature = env->GetMethodID(pmClass, "hasSystemFeature",
+                                            "(Ljava/lang/String;)Z");
+    if (JniCheck::Pending(env) || hasFeature == NULL) {
+        env->DeleteLocalRef(pmClass);
+        env->DeleteLocalRef(pm);
+        env->DeleteLocalRef(activityClass);
+        return 0;
+    }
+
+    jstring feature = env->NewStringUTF("android.software.leanback");
+    if (JniCheck::Pending(env) || feature == NULL) {
+        env->DeleteLocalRef(pmClass);
+        env->DeleteLocalRef(pm);
+        env->DeleteLocalRef(activityClass);
+        return 0;
+    }
+
+    S32 isTV = (S32)env->CallBooleanMethod(pm, hasFeature, feature);
+    JniCheck::Pending(env);
+
+    env->DeleteLocalRef(feature);
+    env->DeleteLocalRef(pmClass);
+    env->DeleteLocalRef(pm);
+    env->DeleteLocalRef(activityClass);
+    return isTV;
 #else
     return 0;
 #endif
