+++ SOURCES/TOUCH_INPUT.CPP
@@ -1,5 +1,6 @@
 #include "TOUCH_INPUT.H"
 
+#include <SYSTEM/ANDROID.H>
 #include <SYSTEM/KEYBOARD.H>
 #include <SYSTEM/KEYBOARD_KEYS.H>
 #include <SYSTEM/LOGPRINT.H>
@@ -8,10 +9,6 @@
 
 #include <SDL3/SDL.h>
 #include <SDL3/SDL_events.h>
-#ifdef __ANDROID__
-#include <jni.h>
-#include <SDL3/SDL_system.h>
-#endif
 
 #include <string.h>
 #include <math.h>
@@ -203,85 +200,6 @@ void ApplyVirtualKeys(void) {
         if (sc < 256 && s_virtual_keys[sc])
             TabKeys[sc] |= 0x80;
     }
-}
-
-/* ---------------------------------------------------------------------------
- * Android TV detection (no touchscreen, skip overlay)
- * --------------------------------------------------------------------------- */
-
-S32 IsAndroidTVDevice(void) {
-#ifdef __ANDROID__
-    struct JniCheck {
-        static S32 Pending(JNIEnv *e) {
-            if (e->ExceptionCheck()) {
-                LogPrintf("TouchOverlay: JNI exception in TV detection, "
-                          "overlay disabled\n");
-                e->ExceptionClear();
-                return 1;
-            }
-            return 0;
-        }
-    };
-    JNIEnv *env = (JNIEnv *)SDL_GetAndroidJNIEnv();
-    if (env == NULL)
-        return 0;
-
-    jobject activity = (jobject)SDL_GetAndroidActivity();
-    if (activity == NULL)
-        return 0;
-
-    jclass activityClass = env->GetObjectClass(activity);
-    if (JniCheck::Pending(env) || activityClass == NULL)
-        return 0;
-
-    jmethodID getPM = env->GetMethodID(activityClass, "getPackageManager",
-                                       "()Landroid/content/pm/PackageManager;");
-    if (JniCheck::Pending(env) || getPM == NULL) {
-        env->DeleteLocalRef(activityClass);
-        return 0;
-    }
-
-    jobject pm = env->CallObjectMethod(activity, getPM);
-    if (JniCheck::Pending(env) || pm == NULL) {
-        env->DeleteLocalRef(activityClass);
-        return 0;
-    }
-
-    jclass pmClass = env->GetObjectClass(pm);
-    if (JniCheck::Pending(env) || pmClass == NULL) {
-        env->DeleteLocalRef(pm);
-        env->DeleteLocalRef(activityClass);
-        return 0;
-    }
-
-    jmethodID hasFeature = env->GetMethodID(pmClass, "hasSystemFeature",
-                                            "(Ljava/lang/String;)Z");
-    if (JniCheck::Pending(env) || hasFeature == NULL) {
-        env->DeleteLocalRef(pmClass);
-        env->DeleteLocalRef(pm);
-        env->DeleteLocalRef(activityClass);
-        return 0;
-    }
-
-    jstring feature = env->NewStringUTF("android.software.leanback");
-    if (JniCheck::Pending(env) || feature == NULL) {
-        env->DeleteLocalRef(pmClass);
-        env->DeleteLocalRef(pm);
-        env->DeleteLocalRef(activityClass);
-        return 0;
-    }
-
-    S32 isTV = (S32)env->CallBooleanMethod(pm, hasFeature, feature);
-    JniCheck::Pending(env);
-
-    env->DeleteLocalRef(feature);
-    env->DeleteLocalRef(pmClass);
-    env->DeleteLocalRef(pm);
-    env->DeleteLocalRef(activityClass);
-    return isTV;
-#else
-    return 0;
-#endif
 }
 
 /* ---------------------------------------------------------------------------
